Add data-dimbox-download-file
Add data attribute data-dimbox-download-file. It defines a file to download by clicking the download button. By default the download button downloads the image/video being displayed. With this attribute it can be overriden for example to download a high resolution version of the image or a any file for example when showing inline content.

Rename function downloadRemoteImage to downloadRemoteFile.